added some moethods to controller

+++ src/main/java/demo/BookRepository.java
@@ -7,6 +7,8 @@ import demo.Book;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called bookRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 
 }
+
+
